@@ -4,44 +4,47 @@
     "description": "https://idg.tools | Inner Development Toolbox",
     "private": true,
     "scripts": {
-        "dev": "svelte-kit dev",
-        "dev:host": "svelte-kit dev --host",
+        "dev": "vite dev",
+        "dev:host": "vite dev --host",
         "build:content": "node scripts/esbuild.js && node scripts/compiled/build-content.js",
-        "build": "svelte-kit build",
+        "build": "vite build",
         "build:all": "npm run build:content && svelte-kit build",
-        "preview": "svelte-kit preview",
+        "preview": "vite preview",
         "check": "svelte-check --tsconfig ./tsconfig.json",
         "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
         "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. .",
         "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
     },
     "dependencies": {
-        "sanitize-html": "^2.7.0",
+        "dompurify": "^2.3.10",
+        "sanitize-html": "^2.7.1",
         "svelte-markdown": "^0.2.2"
     },
     "devDependencies": {
-        "@sveltejs/adapter-vercel": "1.0.0-next.58",
-        "@sveltejs/kit": "^1.0.0-next.350",
-        "@tailwindcss/typography": "^0.5.2",
+        "@sveltejs/adapter-vercel": "1.0.0-next.66",
+        "@sveltejs/kit": "^1.0.0-next.401",
+        "@tailwindcss/typography": "^0.5.4",
+        "@types/dompurify": "^2.3.3",
         "@types/sanitize-html": "^2.6.2",
-        "autoprefixer": "^10.4.7",
-        "cssnano": "^5.1.11",
+        "autoprefixer": "^10.4.8",
+        "cssnano": "^5.1.12",
         "dotenv": "^16.0.1",
         "native-fetch": "^4.0.2",
         "postcss": "^8.4.14",
-        "postcss-load-config": "^3.1.4",
+        "postcss-load-config": "^4.0.1",
         "prettier": "^2.7.1",
         "prettier-plugin-svelte": "^2.7.0",
-        "prettier-plugin-tailwindcss": "^0.1.11",
+        "prettier-plugin-tailwindcss": "^0.1.13",
         "remark-slate": "^1.8.6",
         "slugify": "^1.6.5",
-        "svelte": "^3.48.0",
-        "svelte-check": "^2.7.2",
+        "svelte": "^3.49.0",
+        "svelte-check": "^2.8.0",
         "svelte-preprocess": "^4.10.7",
-        "tailwindcss": "^3.1.3",
+        "tailwindcss": "^3.1.7",
         "tslib": "^2.4.0",
-        "typescript": "^4.7.3",
-        "undici": "^5.5.1"
+        "typescript": "^4.7.4",
+        "undici": "^5.8.0",
+        "vite": "^3.0.4"
     },
     "type": "module",
     "prettier": {
