@@ -1,7 +1,7 @@
 import { content } from '$lib/content-backend'
 
 /** @type {import('@sveltejs/kit').RequestHandler} */
-export async function get() {
+export async function GET() {
     if (content) {
         return { body: { content } }
     }
